Fixed some bugs on the Base page - OutUp v0.0.2

+++ client/src/app/base/base.component.ts
@@ -1,7 +1,6 @@
-import { rendererTypeName } from '@angular/compiler';
-import { Component, OnInit } from '@angular/core';
+import { Component, ElementRef, OnDestroy, OnInit, Renderer2, ViewChild } from '@angular/core';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
-import { Observable } from 'rxjs';
+import { Observable, Subscription } from 'rxjs';
 import { MaterialService } from '../shared/classes/material.service';
 import { Base } from '../shared/interfaces';
 import { BaseService } from '../shared/services/base.service';
@@ -10,14 +9,18 @@ import { BaseService } from '../shared/services/base.service';
   selector: 'app-base',
   templateUrl: './base.component.html'
 })
-export class BaseComponent implements OnInit {
+export class BaseComponent implements OnInit, OnDestroy {
   form: FormGroup;
   video: File;
   videoLoading: boolean = false;
   videoPreview: string | ArrayBuffer;
   base$: Observable<Base[]>;
+  bSub: Subscription;
+  bbSub: Subscription;
 
-  constructor(private _baseService: BaseService) { }
+  @ViewChild('inputRef') inputRef: ElementRef;
+
+  constructor(private _baseService: BaseService, private _renderer: Renderer2) { }
 
   ngOnInit(): void {
     this.form = new FormGroup({
@@ -25,6 +28,14 @@ export class BaseComponent implements OnInit {
       video: new FormControl('', Validators.required)
     });
     this.base$ = this._baseService.getAll();
+  }
+
+  ngOnDestroy(): void {
+    if (this.bSub) {
+      this.bSub.unsubscribe();
+    } else if (this.bbSub) {
+      this.bbSub.unsubscribe();
+    }
   }
 
   onFileSelect(event: any): void {
@@ -46,14 +57,16 @@ export class BaseComponent implements OnInit {
     reader.readAsDataURL(file);
   }
 
-  onSubmit(): void{
+  onSubmit(): void {
     this.form.disable();
-    this._baseService.create(this.form.value.name, this.video).subscribe(
+    this.bSub = this._baseService.create(this.form.value.name, this.video).subscribe(
       base => {
         MaterialService.toast('Exercise added!');
-        this.videoPreview = "";
+        this.inputRef.nativeElement.text = null;
+        this._renderer.setProperty(this.inputRef.nativeElement, 'value', '');
+        this.videoPreview = '';
         this.form.enable();
-      }, 
+      },
       error => {
         MaterialService.toast(error.error.message);
         this.form.enable();
@@ -66,10 +79,19 @@ export class BaseComponent implements OnInit {
   }
 
   deleteExercise(id: string): void {
-    this._baseService.delete(id).subscribe(
-      response => MaterialService.toast(response.message),
-      error => MaterialService.toast(error.error.message),
-      () => this.base$ = this._baseService.getAll()
+    this.bbSub = this._baseService.getExercise(id).subscribe(
+      base => {
+        if (base.use > 0) {
+          MaterialService.toast("You cannot delete this exercise because it is in use!")
+        } else {
+          this._baseService.delete(id).subscribe(
+            response => MaterialService.toast(response.message),
+            error => MaterialService.toast(error.error.message),
+            () => this.base$ = this._baseService.getAll()
+          );
+        }
+      },
+      error => MaterialService.toast(error.error.message)
     );
   }
 }
